test: split out ads
