+++ packages/markup/__tests__/android/markup-ad.android.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared-ad.native";
+
+describe("android", () => {
+  shared();
+});
