+++ packages/markup/__tests__/ios/markup-ad.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared-ad.native";
+
+describe("ios", () => {
+  shared();
+});
