+++ packages/markup/__tests__/shared-ad.native.js
@@ -3,7 +3,7 @@ import {
   minimalNative
 } from "@times-components/jest-serializer";
 import TestRenderer from "react-test-renderer";
-import shared from "./shared.base";
+import shared from "./shared-ad.base";
 
 jest.mock("@times-components/ad", () => "Ad");
 jest.mock("@times-components/pull-quote", () => "PullQuote");
