+++ packages/markup/__tests__/web/markup-ad.web.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared-ad.web";
+
+describe("web", () => {
+  shared();
+});
